requirements.txt
google-generativeai
python-dotenv
yt-dlp